refactored code into three classes like meteor has done it
